Fix for Live 12.1b browser functionality
Optimized the openplugin action to execute "enter, enter" instead of "down, enter" when selecting/opening a menu item. This makes the menu compatible with Live 12.1b and should be backwards compatible with older versions of Live. Also added a function to automatically close the live browser after a menu item has been opened.